+++ Modules/Microsoft365DSC/DSCResources/MSFT_PPTenantIsolationSettings/MSFT_PPTenantIsolationSettings.psm1
@@ -405,8 +405,13 @@ function Set-TargetResource
         [Array]$newRules = $existingAllowedRules | Where-Object -FilterScript { $_.tenantId -notin $removeRules }
         $tenantIsolationPolicy.Properties.allowedTenants = $newRules
     }
+    $tenantGuid = [System.Guid]::Empty
+    if (-not [System.Guid]::TryParse($TenantId, [ref]$tenantGuid))
+    {
+        $TenantId = Get-M365TenantId -TenantName $TenantId
+    }
     $uri = "https://" + (Get-MSCloudLoginConnectionProfile -Workload 'PowerPlatformREST').BapEndpoint + `
-               "/providers/PowerPlatform.Governance/v1/tenants/$($tenantId)/tenantIsolationPolicy?api-version=2020-06-01"
+               "/providers/PowerPlatform.Governance/v1/tenants/$($TenantId)/tenantIsolationPolicy?api-version=2020-06-01"
     Write-Verbose -Message "Updating with payload:`r`n$(ConvertTo-Json $tenantIsolationPolicy -Depth 20)"
     Invoke-M365DSCPowerPlatformRESTWebRequest -Uri $uri -Method 'PUT' -Body $tenantIsolationPolicy
 }
@@ -739,6 +744,7 @@ function Export-TargetResource
     }
 }
 
+$Script:TenantNameToGuidCache = @{}
 function Get-M365TenantId
 {
     param
@@ -753,9 +759,16 @@ function Get-M365TenantId
         return '*'
     }
 
+    if ($Script:TenantNameToGuidCache.ContainsKey($TenantName))
+    {
+        return $Script:TenantNameToGuidCache[$TenantName]
+    }
+
     $result = Invoke-WebRequest "https://login.windows.net/$TenantName/.well-known/openid-configuration" -UseBasicParsing
     $jsonResult = $result | ConvertFrom-Json
-    return $jsonResult.token_endpoint.Split('/')[3]
+    $tenantId = $jsonResult.token_endpoint.Split('/')[3]
+    $Script:TenantNameToGuidCache[$TenantName] = $tenantId
+    return $tenantId
 }
 
 Export-ModuleMember -Function *-TargetResource